+++ resources/views/banking/transfers/index.blade.php
@@ -3,15 +3,15 @@
 @section('title', trans_choice('general.transfers', 2))
 
 @section('new_button')
-    @permission('create-banking-transfers')
-            <a href="{{ route('transfers.create') }}" class="btn btn-success btn-sm header-button-top"><span class="fa fa-plus"></span> &nbsp;{{ trans('general.add_new') }}</a>
-    @endpermission
-    <span><a href="{{ route('import.create', ['banking', 'transfers']) }}" class="btn btn-white btn-sm header-button-top"><span class="fa fa-upload "></span> &nbsp;{{ trans('import.import') }}</a></span>
-    <span><a href="{{ route('transfers.export', request()->input()) }}" class="btn btn-white btn-sm header-button-top"><span class="fa fa-download"></span> &nbsp;{{ trans('general.export') }}</a></span>
+    @can('create-banking-transfers')
+        <a href="{{ route('transfers.create') }}" class="btn btn-success btn-sm">{{ trans('general.add_new') }}</a>
+    @endcan
+    <a href="{{ route('import.create', ['banking', 'transfers']) }}" class="btn btn-white btn-sm">{{ trans('import.import') }}</a>
+    <a href="{{ route('transfers.export', request()->input()) }}" class="btn btn-white btn-sm">{{ trans('general.export') }}</a>
 @endsection
 
 @section('content')
-    @if ($transfers->count())
+    @if ($transfers->count() || request()->get('search', false))
         <div class="card">
             <div class="card-header border-bottom-0" :class="[{'bg-gradient-primary': bulk_action.show}]">
                 {!! Form::open([
@@ -21,10 +21,7 @@
                     'class' => 'mb-0'
                 ]) !!}
                     <div class="align-items-center" v-if="!bulk_action.show">
-                        <akaunting-search
-                            :placeholder="'{{ trans('general.search_placeholder') }}'"
-                            :options="{{ json_encode([]) }}"
-                        ></akaunting-search>
+                        <x-search-string model="App\Models\Banking\Transfer" />
                     </div>
 
                     {{ Form::bulkActionRowGroup('general.transfers', $bulk_actions, ['group' => 'banking', 'type' => 'transfers']) }}
@@ -46,12 +43,19 @@
 
                     <tbody>
                         @foreach($transfers as $item)
+                            @php
+                            $item->name = trans('transfers.messages.delete', [
+                                'from' => $item->expense_transaction->account->name,
+                                'to' => $item->income_transaction->account->name,
+                                'amount' => money($item->expense_transaction->amount, $item->expense_transaction->currency_code, true)
+                            ]);
+                            @endphp
                             <tr class="row align-items-center border-top-1">
-                                <td class="col-sm-2 col-md-1 d-none d-sm-block">{{ Form::bulkActionGroup($item->id, $item->from_account) }}</td>
-                                <td class="col-md-2 d-none d-md-block"><a class="col-aka" href="{{ route('transfers.edit', $item->id) }}">@date($item->paid_at)</a></td>
-                                <td class="col-sm-2 col-md-3 d-none d-sm-block">{{ $item->from_account }}</td>
-                                <td class="col-xs-4 col-sm-4 col-md-2">{{ $item->to_account }}</td>
-                                <td class="col-xs-4 col-sm-2 col-md-2 text-right">@money($item->amount, $item->currency_code, true)</td>
+                                <td class="col-sm-2 col-md-1 d-none d-sm-block">{{ Form::bulkActionGroup($item->id, $item->expense_transaction->account->name) }}</td>
+                                <td class="col-md-2 d-none d-md-block"><a class="col-aka" href="{{ route('transfers.edit', $item->id) }}">@date($item->expense_transaction->paid_at)</a></td>
+                                <td class="col-sm-2 col-md-3 d-none d-sm-block">{{ $item->expense_transaction->account->name }}</td>
+                                <td class="col-xs-4 col-sm-4 col-md-2">{{ $item->income_transaction->account->name }}</td>
+                                <td class="col-xs-4 col-sm-2 col-md-2 text-right">@money($item->expense_transaction->amount, $item->expense_transaction->currency_code, true)</td>
                                 <td class="col-xs-4 col-sm-2 col-md-2 text-center">
                                     <div class="dropdown">
                                         <a class="btn btn-neutral btn-sm text-light items-align-center py-2" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -59,10 +63,10 @@
                                         </a>
                                         <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
                                             <a class="dropdown-item" href="{{ route('transfers.edit', $item->id) }}">{{ trans('general.edit') }}</a>
-                                            @permission('delete-banking-transfers')
+                                            @can('delete-banking-transfers')
                                                 <div class="dropdown-divider"></div>
                                                 {!! Form::deleteLink($item, 'transfers.destroy') !!}
-                                            @endpermission
+                                            @endcan
                                         </div>
                                     </div>
                                 </td>
@@ -79,7 +83,7 @@
             </div>
         </div>
     @else
-        @include('partials.admin.empty_page', ['page' => 'transfers', 'docs_path' => 'banking/transfers'])
+        <x-empty-page group="banking" page="transfers" />
     @endif
 @endsection
