+++ resources/views/wizard/currencies/index.blade.php
@@ -68,10 +68,10 @@
                                             <button type="button" class="dropdown-item" @click="onEditCurrency('{{ $item->id }}')">
                                                 {{ trans('general.edit') }}
                                             </button>
-                                            @permission('delete-settings-currencies')
+                                            @can('delete-settings-currencies')
                                                 <div class="dropdown-divider"></div>
                                                 {!! Form::deleteLink($item, 'wizard/currencies') !!}
-                                            @endpermission
+                                            @endcan
                                         </div>
                                     </div>
                                 </td>
@@ -121,13 +121,13 @@
         <div class="card-footer">
             <div class="row">
                 <div class="col-sm-6">
-                    <a href="{{ route('wizard.companies.edit') }}" class="btn btn-icon btn-white header-button-top">
+                    <a href="{{ route('wizard.companies.edit') }}" class="btn btn-icon btn-white">
                         <span class="btn-inner--text">{{ trans('pagination.previous') }}</span>
                     </a>
                 </div>
 
                 <div class="col-sm-6 text-right">
-                    <a href="{{ route('wizard.taxes.index') }}" id="wizard-skip" class="btn btn-icon btn-white header-button-top">
+                    <a href="{{ route('wizard.taxes.index') }}" id="wizard-skip" class="btn btn-icon btn-white">
                         <span class="btn-inner--text">{{ trans('pagination.next') }}</span>
                     </a>
                 </div>
